- Temporary commit to save work to make reference only nuget packages.
This causes the DLLs to not be copied into the bin folder and also ensures they do not make their way to other projects referencing the plugin project. But it means we wouldn't be able to use this NuGet packages in Rock Mobile. In the future we might use this feature to make *.Ref packages.